# services/api/openapi.json
{
  "components": {
    "schemas": {
      "ApiEnvelope": {
        "additionalProperties": false,
        "description": "Consistent success wrapper. Domain meaning stays inside result.",
        "properties": {
          "api_version": {
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "execution_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Status"
          },
          "input_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Hash"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "product_version": {
            "default": "0.24.3-19",
            "title": "Product Version",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "scientific_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scientific Status"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "request_id",
          "result"
        ],
        "title": "ApiEnvelope",
        "type": "object"
      },
...
      "CompareEvidenceRequest": {
        "additionalProperties": false,
        "description": "Evidence mode reuses two identified variant identities.",
        "properties": {
          "assembly_a": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Assembly A",
            "type": "string"
          },
          "assembly_b": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Assembly B",
            "type": "string"
          },
          "text_a": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text A",
            "type": "string"
          },
          "text_b": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text B",
            "type": "string"
          }
        },
        "required": [
          "text_a",
          "assembly_a",
          "text_b",
          "assembly_b"
        ],
        "title": "CompareEvidenceRequest",
        "type": "object"
      },
      "CompareEvolutionRequest": {
        "additionalProperties": false,
        "properties": {
          "column": {
            "default": 0,
            "minimum": 0.0,
            "title": "Column",
            "type": "integer"
          },
          "fasta": {
            "maxLength": 2000000,
            "minLength": 1,
            "title": "Fasta",
            "type": "string"
          },
          "group_ids": {
            "items": {
              "type": "string"
            },
            "title": "Group Ids",
            "type": "array"
          },
          "member_id": {
            "default": "",
            "maxLength": 80,
            "title": "Member Id",
            "type": "string"
          }
        },
        "required": [
          "fasta"
        ],
        "title": "CompareEvolutionRequest",
        "type": "object"
      },
      "CompareGuidesRequest": {
        "additionalProperties": false,
        "properties": {
          "guide_a": {
            "additionalProperties": true,
            "title": "Guide A",
            "type": "object"
          },
          "guide_b": {
            "additionalProperties": true,
            "title": "Guide B",
            "type": "object"
          }
        },
        "required": [
          "guide_a",
          "guide_b"
        ],
        "title": "CompareGuidesRequest",
        "type": "object"
      },
      "CompareParseRequest": {
        "additionalProperties": false,
        "properties": {
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "payload"
        ],
        "title": "CompareParseRequest",
        "type": "object"
      },
      "CompareProteinsRequest": {
        "additionalProperties": false,
        "properties": {
          "identifier_a": {
            "default": "protein_a",
            "maxLength": 80,
            "title": "Identifier A",
            "type": "string"
          },
          "identifier_b": {
            "default": "protein_b",
            "maxLength": 80,
            "title": "Identifier B",
            "type": "string"
          },
          "sequence_a": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Sequence A",
            "type": "string"
          },
          "sequence_b": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Sequence B",
            "type": "string"
          }
        },
        "required": [
          "sequence_a",
          "sequence_b"
        ],
        "title": "CompareProteinsRequest",
        "type": "object"
      },
      "CompareRemoteJobRequest": {
        "additionalProperties": false,
        "properties": {
          "fetch_coordinates": {
            "default": true,
            "title": "Fetch Coordinates",
            "type": "boolean"
          },
          "method": {
            "default": "tm-align",
            "maxLength": 40,
            "title": "Method",
            "type": "string"
          },
          "reference_chain": {
            "default": "A",
            "maxLength": 8,
            "title": "Reference Chain",
            "type": "string"
          },
          "reference_entry": {
            "maxLength": 80,
            "minLength": 4,
            "pattern": "^[0-9A-Za-z_-]+$",
            "title": "Reference Entry",
            "type": "string"
          },
          "target_chain": {
            "default": "A",
            "maxLength": 8,
            "title": "Target Chain",
            "type": "string"
          },
          "target_entry": {
            "maxLength": 80,
            "minLength": 4,
            "pattern": "^[0-9A-Za-z_-]+$",
            "title": "Target Entry",
            "type": "string"
          }
        },
        "required": [
          "reference_entry",
          "target_entry"
        ],
        "title": "CompareRemoteJobRequest",
        "type": "object"
      },
      "CompareUSalignJobRequest": {
        "additionalProperties": false,
        "properties": {
          "reference_chain": {
            "default": "A",
            "maxLength": 8,
            "title": "Reference Chain",
            "type": "string"
          },
          "reference_entry": {
            "pattern": "^(1CRN|1BNA|1RNA)$",
            "title": "Reference Entry",
            "type": "string"
          },
          "target_chain": {
            "default": "A",
            "maxLength": 8,
            "title": "Target Chain",
            "type": "string"
          },
          "target_entry": {
            "pattern": "^(1CRN|1BNA|1RNA)$",
            "title": "Target Entry",
            "type": "string"
          }
        },
        "required": [
          "reference_entry",
          "target_entry"
        ],
        "title": "CompareUSalignJobRequest",
        "type": "object"
      },
      "CompareVariantsRequest": {
        "additionalProperties": false,
        "properties": {
          "assembly_a": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Assembly A",
            "type": "string"
          },
          "assembly_b": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Assembly B",
            "type": "string"
          },
          "text_a": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text A",
            "type": "string"
          },
          "text_b": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Text B",
            "type": "string"
          }
        },
        "required": [
          "text_a",
          "assembly_a",
          "text_b",
          "assembly_b"
        ],
        "title": "CompareVariantsRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobAccepted": {
        "properties": {
          "execution_status": {
            "title": "Execution Status",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "poll_url": {
            "title": "Poll Url",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "job_id",
          "kind",
          "execution_status",
          "poll_url"
        ],
        "title": "JobAccepted",
        "type": "object"
      },
...
      "ReferenceAdmissionRequest": {
        "additionalProperties": false,
        "properties": {
          "assembly_id": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Assembly Id",
            "type": "string"
          }
        },
        "required": [
          "assembly_id"
        ],
        "title": "ReferenceAdmissionRequest",
        "type": "object"
      },
      "ReferenceInstallRequest": {
        "additionalProperties": false,
        "properties": {
          "assembly_id": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Assembly Id",
            "type": "string"
          },
          "confirm": {
            "default": false,
            "title": "Confirm",
            "type": "boolean"
          }
        },
        "required": [
          "assembly_id"
        ],
        "title": "ReferenceInstallRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Product version 0.24.3-19. API contract v1. This service validates transport input, calls helixscope_core, and serializes results. It does not independently calculate biology. Job backend: LOCAL MIGRATION JOB BACKEND, NON-DURABLE, NOT FINAL PRODUCTION QUEUE.",
    "summary": "Typed scientific HTTP API for HelixScope core.",
    "title": "HelixScope API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/compare/evidence": {
      "post": {
        "description": "Does not vote, rank, or diagnose.",
        "operationId": "compareEvidence",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evidence pack for two identified variants",
        "tags": [
          "compare"
        ]
      }
    },
    "/api/v1/compare/evolution": {
      "post": {
        "description": "Conservation among submitted sequences. Not dN/dS.",
        "operationId": "compareEvolution",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareEvolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Inspect MSA columns and optional group-associated positions",
        "tags": [
          "compare"
        ]
      }
    },
    "/api/v1/compare/guides": {
      "post": {
        "description": "No invented universal score. Genome-wide MIT/CFD stay unavailable unless supplied.",
        "operationId": "compareGuides",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareGuidesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare two CRISPR guides",
        "tags": [
          "compare"
        ]
      }
    },
    "/api/v1/compare/jobs/remote": {
      "post": {
        "description": "Explicit RCSB backend. Does not fall back to US-align or rewrite IDs/chains/method.",
        "operationId": "compareRemoteSubmitJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRemoteJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue RCSB Alignment API comparison",
        "tags": [
          "compare"
        ]
      }
    },
    "/api/v1/compare/jobs/usalign": {
      "post": {
        "description": "Explicit local US-align backend. Does not fall back to RCSB or substitute 1CRN/1BNA.",
        "operationId": "compareUSalignSubmitJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareUSalignJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Queue local US-align on bundled structures",
        "tags": [
          "compare"
        ]
      }
    },
...
    "/api/v1/compare/parse": {
      "post": {
        "description": "Keeps block RMSD and global RMSD as distinct fields.",
        "operationId": "compareParseAlignment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Parse an RCSB COMPLETE alignment payload",
        "tags": [
          "compare"
        ]
      }
    },
    "/api/v1/compare/proteins": {
      "post": {
        "description": "Sequence identity is not structure similarity.",
        "operationId": "compareProteins",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareProteinsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare two protein sequences",
        "tags": [
          "compare"
        ]
      }
    },
    "/api/v1/compare/variants": {
      "post": {
        "description": "Identity side-by-side. ClinVar/VEP layers are not invented here.",
        "operationId": "compareVariants",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareVariantsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare two identified variants without pathogenicity ranking",
        "tags": [
          "compare"
        ]
      }
    },
...
    "/api/v1/references/admission": {
      "post": {
        "description": "Does not download. TEST REFERENCE is not a public genome.",
        "operationId": "referencesAdmitDownload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceAdmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resource admission for a catalogued public assembly",
        "tags": [
          "references"
        ]
      }
    },
    "/api/v1/references/install": {
      "post": {
        "description": "Explicit confirm required. RESOURCE_LIMIT is not bypassed. No arbitrary URL.",
        "operationId": "referencesInstallCatalog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceInstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a catalogued reference download after admission PROCEED",
        "tags": [
          "references"
        ]
      }
    },
...
  }
}